[ADAM-634] Remove bdg-utils-parquet
