+++ bdg-utils-io/src/main/scala/org/bdgenomics/utils/io/ByteArrayByteAccess.scala
@@ -1,0 +1,35 @@
+/**
+ * Licensed to Big Data Genomics (BDG) under one
+ * or more contributor license agreements.  See the NOTICE file
+ * distributed with this work for additional information
+ * regarding copyright ownership.  The BDG licenses this file
+ * to you under the Apache License, Version 2.0 (the
+ * "License"); you may not use this file except in compliance
+ * with the License.  You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package org.bdgenomics.utils.io
+
+import java.io.{ ByteArrayInputStream, InputStream }
+
+class ByteArrayByteAccess(val bytes: Array[Byte]) extends ByteAccess with Serializable {
+
+  private val inputStream = new ByteArrayInputStream(bytes)
+  assert(inputStream.markSupported(), "ByteArrayInputStream doesn't support marks")
+
+  inputStream.mark(bytes.length)
+
+  override def length(): Long = bytes.length
+  override def readByteStream(offset: Long, length: Int): InputStream = {
+    inputStream.reset()
+    inputStream.skip(offset)
+    inputStream
+  }
+}
