+++ bdg-utils-io/src/main/scala/org/bdgenomics/utils/io/S3ByteAccess.scala
@@ -1,0 +1,36 @@
+/**
+ * Licensed to Big Data Genomics (BDG) under one
+ * or more contributor license agreements.  See the NOTICE file
+ * distributed with this work for additional information
+ * regarding copyright ownership.  The BDG licenses this file
+ * to you under the Apache License, Version 2.0 (the
+ * "License"); you may not use this file except in compliance
+ * with the License.  You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package org.bdgenomics.utils.io
+
+import java.io.InputStream
+
+import com.amazonaws.services.s3.AmazonS3
+import com.amazonaws.services.s3.model.GetObjectRequest
+
+class S3ByteAccess(client: AmazonS3, bucket: String, keyName: String) extends ByteAccess {
+  assert(bucket != null)
+  assert(keyName != null)
+
+  lazy val objectMetadata = client.getObjectMetadata(bucket, keyName)
+  override def length(): Long = objectMetadata.getContentLength
+  override def readByteStream(offset: Long, length: Int): InputStream = {
+    val getObjectRequest = new GetObjectRequest(bucket, keyName).withRange(offset, offset + length)
+    client.getObject(getObjectRequest).getObjectContent
+  }
+
+}
