+++ pom.xml
@@ -31,7 +31,7 @@
     <module>bdg-utils-misc</module>
     <module>bdg-utils-minhash</module>
     <module>bdg-utils-metrics</module>
-    <module>bdg-utils-parquet</module>
+    <module>bdg-utils-io</module>
     <module>bdg-utils-statistics</module>
     <module>bdg-utils-cli</module>
     <module>bdg-utils-serialization</module>
